% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation.R
\name{annotate_protein_coding}
\alias{annotate_protein_coding}
\title{Annotate Protein-Coding Genes}
\usage{
annotate_protein_coding(data_path, gtf_path, project_name = "scRNA_analysis")
}
\arguments{
\item{data_path}{Character. Path to the single-cell dataset (H5 format)}

\item{gtf_path}{Character. Path to the GTF annotation file}

\item{project_name}{Character. Name for the Seurat project (default: "scRNA_analysis")}
}
\value{
A Seurat object containing only protein-coding genes
}
\description{
Identifies and filters protein-coding genes from GTF annotation file and
retains only these genes in the Seurat object.
}
\examples{
\dontrun{
seurat_obj <- annotate_protein_coding(
  data_path = "data/filtered_feature_bc_matrix",
  gtf_path = "data/Homo_sapiens.GRCh38.111.gtf"
)
}

}
