% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes.R
\name{filter_protein_coding_genes}
\alias{filter_protein_coding_genes}
\title{Filter protein-coding genes from GTF}
\usage{
filter_protein_coding_genes(gtf)
}
\arguments{
\item{gtf}{A GRanges object imported from a GTF file.}
}
\value{
A filtered GRanges object containing only protein-coding genes.
}
\description{
This function filters a GTF GRanges object to retain only entries that are
annotated as protein-coding genes.
}
