% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{perform_optimized_pca}
\alias{perform_optimized_pca}
\title{Perform PCA with memory optimization}
\usage{
perform_optimized_pca(seurat_obj, npcs = 50, verbose = TRUE)
}
\arguments{
\item{seurat_obj}{Seurat object (normalized)}

\item{npcs}{Number of principal components to compute}

\item{verbose}{Logical, show progress}
}
\value{
Seurat object with PCA computed
}
\description{
Perform PCA with memory optimization
}
