% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_analysis.R
\name{perform_pca_analysis}
\alias{perform_pca_analysis}
\title{Perform PCA Analysis and Plot Variance Explained}
\usage{
perform_pca_analysis(
  seurat_obj,
  n_pcs = 20,
  save_path_rds = NULL,
  save_plot = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{n_pcs}{Integer. Number of principal components to compute (default: 20)}

\item{save_path_rds}{Path to save the Seurat object after PCA (default: NULL)}

\item{save_plot}{Path to save PCA variance histogram as PNG (default: NULL)}
}
\value{
A Seurat object with PCA computed
}
\description{
Normalizes, scales and performs PCA on a Seurat object. Saves PCA variance plot.
}
\examples{
\dontrun{
seurat_obj <- perform_pca_analysis(
  seurat_obj,
  n_pcs = 20,
  save_path_rds = "data/seurat_obj.pca.rds",
  save_plot = "data/pca_variance_histogram.png"
)
}

}
