% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_plot.R
\name{plot_genes_expressed_violin}
\alias{plot_genes_expressed_violin}
\title{Plot number of genes expressed per cell from previous function}
\usage{
plot_genes_expressed_violin(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object with the metadata column `genes_expr_3umis`.}
}
\value{
A ggplot object showing the violin plot.
}
\description{
Plots a violin plot of the number of genes (≥3 UMIs) per cell to visualize result of previous function.
take the ggplot2 library, create a data frame for cells and genes expressed and plot the violin plot.
}
\examples{
plot_genes_expressed_violin(seurat_obj)
}
