% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_PCA.R
\name{run_basic_pca}
\alias{run_basic_pca}
\title{Run basic PCA on a Seurat object}
\usage{
run_basic_pca(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object containing raw count data}
}
\value{
A Seurat object updated with PCA results
}
\description{
This function normalizes the data, identifies variable genes, scales the data,
and performs Principal Component Analysis (PCA) on the dataset.
}
\examples{
tu use the function and visualize the PCA result:
seurat_obj <- run_basic_pca(seurat_obj)
print(seurat_obj[["pca"]])
ElbowPlot(seurat_obj)
}
