% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_UMAP_and_clustering.R
\name{run_umap_and_cluster}
\alias{run_umap_and_cluster}
\title{Run UMAP and clustering on Seurat object}
\usage{
run_umap_and_cluster(seurat_obj, dims = 10)
}
\arguments{
\item{seurat_obj}{A Seurat object with PCA results already computed.}

\item{dims}{Number of principal components to use (10).}
}
\value{
The Seurat object with UMAP and clustering results added.
}
\description{
This function performs dimensionality reduction and clustering on a Seurat object (with PCA already ran)
using the specified number of principal components (10, the most informative). method to visualize in 2D the PCA result.
}
\examples{
seurat_obj <- run_umap_and_cluster(seurat_obj, dims = 10)
DimPlot(seurat_obj, reduction = "umap", label = TRUE) +
ggplot2::ggtitle("UMAP of clustered cells")
}
