% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAanalysis-package.R
\docType{package}
\name{scRNAseqTools-package}
\alias{scRNAseqTools}
\alias{scRNAseqTools-package}
\title{scRNAseqTools: Single Cell RNA-seq Analysis Tools}
\description{
The implemented functions guide users through a complete analysis workflow, from gene annotation to biological interpretation
}
\author{
\strong{Maintainer}: Giorgia Gentile \email{gentilegiorgia01@gmail.com}

}
\keyword{internal}
