% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_removed_genes.R
\name{summarize_removed_genes}
\alias{summarize_removed_genes}
\title{Summarize number of unwanted genes by category}
\usage{
summarize_removed_genes(gtf_pc)
}
\arguments{
\item{gtf_pc}{A GRanges object containing protein-coding genes (before filtering)}
}
\value{
A data.frame with the number of removed genes per category
}
\description{
Creates a summary table of how many mitochondrial genes, ribosomal genes, and pseudogenes
were identified in the GRanges object before filtering.gives amount og removed genes from each category.
}
\examples{
to use the function:
summary_table <- summarize_removed_genes(gtf_pc)
print(summary_table)
}
